third_party/blink/web_tests/fast/dom/shadow/resize-in-shadow-dom-expected.txt

PASS box.offsetWidth is 114
PASS box.offsetHeight is 114
PASS successfullyParsed is true

TEST COMPLETE
